+++ CHANGELOG.md
@@ -12,9 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - `AppService` is replaced by four smaller, more specialized services
-  - `SessionService` - knows (but does not tell) where to save session tokens, and knows whether the user has an active one
+  - `SessionService` - knows (but does not tell) where to save session tokens; knows whether the user has an active one; and allows to fetch (and cache) authorization details
   - `AuthenticationService` - allows to login as user (or guest) and register, and to cancel any ongoing instance of these processes
-  - `AuthorizationService` - allows to fetch (and cache) authorized access to APIs. Also, routinely checks validity of session token once an authentication is made
   - `ProfileService` - allows to fetch (and cache) the current user's profile data, and to update it as well
 - `ManagementHeaderComponent` refactored down to simpler components
   - Introduced a common `HeaderBrandComponent` used in both store and management pages